omplasserte ekspandere, begynte på statistikk og litt opprydding
